# Report whether video content is optimized for streaming.