--- a/avcodec.go
+++ b/avcodec.go
@@ -13,10 +13,15 @@ package lilliput
 import "C"
 
 import (
+	"bytes"
+	"encoding/binary"
 	"io"
 	"time"
 	"unsafe"
 )
+
+const probeBytesLimit = 32 * 1024
+const atomHeaderSize = 8
 
 type avCodecDecoder struct {
 	decoder    C.avcodec_decoder
@@ -56,6 +61,51 @@ func (d *avCodecDecoder) Description() string {
 	}
 
 	return fmt
+}
+
+func (d *avCodecDecoder) IsStreamable() bool {
+	var atomSize uint32
+	var atomType [4]byte
+	bytesRead := int64(0)
+
+	reader := bytes.NewReader(d.buf)
+
+	// Read atoms within the probe limit
+	for bytesRead < probeBytesLimit && bytesRead < int64(len(d.buf)) {
+		// Read atom header
+		err := binary.Read(reader, binary.BigEndian, &atomSize)
+		if err != nil {
+			break
+		}
+		err = binary.Read(reader, binary.BigEndian, &atomType)
+		if err != nil {
+			break
+		}
+		bytesRead += int64(atomHeaderSize)
+
+		// Check for 'moov' atom
+		if string(atomType[:]) == "moov" {
+			return true
+		}
+
+		if string(atomType[:]) == "mdat" {
+			// 'mdat' encountered before 'moov'
+			return false
+		}
+
+		// Skip to next atom (if the atom is not 'moov' or 'mdat')
+		nextAtomPosition := int64(atomSize) - atomHeaderSize
+		if bytesRead+nextAtomPosition > probeBytesLimit {
+			// If the next atom position exceeds the probe limit or the length of data, stop
+			break
+		}
+
+		// Move the reader to the next atom position
+		reader.Seek(nextAtomPosition, 1)
+		bytesRead += nextAtomPosition
+	}
+
+	return false
 }
 
 func (d *avCodecDecoder) Duration() time.Duration {